Support multiple Play json versions concurrently
These are the consumers of this library:
* frontend: Scala 2.13, Play 2.8 & 3.0
* MAPI: Scala 2.13, Play 2.8
* targeting: Scala 2.13, Play 2.8

We now want to upgrade frontend to Play 3: https://github.com/guardian/frontend/pull/26755 but we also need to ensures that versions of this library are available to consumers that are stil on Play 2.8.

The updated sbt configuration is based on techniques used in https://github.com/guardian/facia-scala-client and https://github.com/guardian/play-googleauth where it's been successful for a few years now.

Co-authored-by: Roberto Tyley <roberto.tyley@guardian.co.uk>